Fix type inversion for shift and exp operators.

+++ test/libsolidity/semanticTests/cleanup/exp_cleanup.sol
@@ -1,5 +1,5 @@
 contract C {
-    function f() public pure returns (uint8 x) {
+    function f() public pure returns (uint x) {
         uint8 y = uint8(2)**uint8(8);
         return 0**y;
     }
